Make IO.relativize relativize less aggressively
Specifically, following the scaladoc of relativize, make sure (first)
that "base" is not a parent of "file" before trying to relativize.

Prior to this change invoking

    IO.relativize(
      file("/target/scala-2.11/resource_managed/main"),   // base
      file("/src/main/resources/scalac-plugin.xml")       // file
    )

returned

    Some("../../../../src/main/resources/scalac-plugin.xml")

which is incorrect according to the scaladoc of relativize and broke
sbt's copyResources implementation.

Created PathMapperSpec to specify and assert the correct behaviour, and
as a regression test.