New `WatchService` based on Java 7's `WatchService`
The goal of this is to add an abstraction layer so that we can define
new kinds of `WatchService`s. This way, we can use the default
`WatchService` provided by the JDK on certain platform (e.g. Windows),
and revert back to the `PollingWatchService` that existed before, on
machine where there's no native support for monitoring the file system
(e.g. macOS).

This commit includes an adapter for the `WatchService` from Java NIO to
our new interface, and a port of the existing `PollingWatchService`.