Tweak WatchState#++ to use the accumulator rather than the initial value
This removes a case of duplicate directory registry, which probably
leads to slower performance.

Just from a code standpoint it doesn't make sense to to be using the
initial ("zero") value of the foldLeft rather than the accumulator.

From a regression point of view this can't cause any regressions because
it's only not double registering files, it doesn't cause some files to
no longer be registered.