Wait 1s before testing for file contents change
Looks like on HFS+ time stamp granularity is only 1 second
https://developer.apple.com/library/content/documentation/FileManagement/Conceptual/APFS_Guide/VolumeFormatComparison/VolumeFormatComparison.html
therefore in order to notice a last modified time change you need to
have elapsed into the next second. We'll do this by sleeping 1 second.

Ref #82

I need to review the downstream issue in sbt to see if #82 should be
closed or not.