Fix redundant registration issue
I noticed that some paths were being repeatedly registered with the
watch service and tracked it down to this code. If a file in the fs
list is not a directory, then its parent would always get registered,
even if there were multiple files with the same parent. The fix is to
use the accumulated map in the fold instead of the seed and to add both
the parent and the child to the accumulated map.