Convert lastModified/setLastModified to Milli calls
Also preserve the current behavior of IO.copyLastModified(),
for compatibility.

The method IO.copyLastModified() has an underspecified, and generally
inconsistent behavior. In particular, if the source file is missing,
it will silently set the target modification time to 1st January 1970,
returning success. Also, because it uses lastModified(), it may
trim away the millisecond part of the timestamp without notice.

The replacement copyModifiedTime(), instead, has true millisecond
precision. Also, it will correctly throw a FileNotFoundException
if any of the two files are missing, or an IOException in case an
IO exception occurs.

This commit deprecates copyLastModified() in favor of the new
copyModifiedTime(), adding a Scaladoc explanation.