Replace utimensat with direct Darwin calls on MacOS
The `utimensat` call is only available in High Sierra, but not
on previous versions of OSX. An alternative could be to use
utimes, but that call always changes both the modification as
well as the last access time, requiring an additional trip
to `stat` in order to reconstruct the necessary information.

A better alternative is just calling the native interface,
similarly to the Windows interface code. This pull request
adds direct support for precise timestamps on the Mac by
using efficient Darwin calls.