WinMilli: enable read/read/delete/rename sharing
In Windows, reading or writing file attributes requires opening
the file via CreateFile(). If only FILE_SHARE_READ is specified
when reading the timestamp, the CreateFile() call will fail
when the file happens to be already open for writing. This
pull request specifies that the file that is being opened has
shared read and write access; the shared delete mode (which also
implies rename) is also allowed, and in case the subsequent
timestamp get/set fails (if the file disappears) the usual
exception-to-zero code path will be taken.

See sbt/sbt#3972