A few changes to the new delete() code
The previous delete() code, which is being replaced by a
new FileVisitor implementation, was tolerant in the case
of deletion of non-existing files or directory. This
change restores the same behavior in the new code.

Also, it renames Path to NioPath to avoid a collision
with sbt.io.Path (see the same renaming in sbt/io/Path.scala),
and applies the standard formatting.