Reduce watch service latency
There is often noticeable lag between when a file event happens and when
the build is triggered. There is also often latency when the user
presses enter to interrupt the watch. Both are due to the way that
SourceModificationWatch polls for results. It can be removed by having
the main thread block on an event queue that is asynchronously filled by
threads that wait for user input and file events. I made a similar
change in CloseWatch.

The new EventMonitor is where most of the logic lives. It is implemented
as a block box that effectively provides three methods:
1) watch() -- block the main thread until a user input or watched file
   event occurs, returning true if the build was triggered
2) state() -- a snapshot of the current watch state. This is primarily
   for legacy compatibility with sbt, which has a number of methods that
   take the state as input. In practice, all that sbt really needs most
   of the time is the count of the number of build triggers
3) close() -- shutdown any threads or service started up by the event
   monitor

I implemented the EventMonitor as a block box so that it would be
straightforward to change the existing implementation or add new
implementations without having to break forward and backwards
compatibility with sbt. In particular, I can envision adding a second
EventMonitor that uses a persistent file system cache for file events
instead of the ad-hoc cache that currently exists in the
EventMonitorImpl.eventThread.

At the moment, the one implementation is EventMonitorImpl. It spins up
two threads for monitoring user input and the file events. Both write
`EventMonitor.Events` to a concurrent queue that the main thread reads
from. User input events supersede file triggers. When the user hits
enter, the queue is cleared and filled with the exit event. Otherwise,
the event thread polls the watch service for events. When it receives a
file event, it adds the event to a cache of recent events. This cache is
used to prevent multiple builds from being triggered by the same file*.
The eventThread also detects when directories are created or deleted and
registers or unregisters the directory with the watch service as needed.

I also stopped registering all of the files. Only directories are
registered now. The registered files just waste memory.

This commit also adds logging of watch events. Unfortunately, I can't
get the logging to actually work at the debug level due to
https://github.com/sbt/sbt/issues/4097, but once that issue is fixed,
logging with the new EventMonitor should work.

I added a simple test that the anti-entropy works and made some small
adjustments to the existing tests to make them work with the new
implementation.

For backwards compatibility with older versions of sbt, I re-implement
SourceModificationWatch.watch to wrap an EventMonitor that we shutdown
after each trigger. This is also the reason that I had to add two close
methods to EventMonitor -- one that shuts down the watch service, and
one that doesn't.

* When neovim, for example, saves a file. It moves the buffer into the
file location which can trigger a delete and a create event on the file.
Without the anti-entropy timeout, there would be two build triggered for
the same file save.