Set file times on register
When registering a new path, we should not trigger events for the files
that already exist in the directory. To fix this, we calculate the last
modified times for all of the files in the newly registered path and add
them to the list.

This fixed spurious test failures that I was having in my anti-entropy
spec.