Add unregister to WatchService
When a watched directroy is deleted, it's good practice to cleanup the
resources associated with it. If it's recreated, we can just
re-register. For binary compatibility reasons, I couldn't add unregister
to WatchService, so I added a trait that gets mixed into all of the
implementations. The method is then accessible in abstract WatchServices
by runtime pattern matching.

This change also has the side effect of preventing the default
WatchService from creating redundant registrations.

Bonus: fix some compiler warnings