Monitor streams
The apple file system api only offers recursive directory listeners,
which makes it pointless to have a listener for each subdirectory.
Whenever a stream is registered with FileEventsApi and there are
previously registered streams that are subdirectories of this stream,
the FileEventsApi removes the now redundant streams and fires a callback
with the name of each directory whose events are now triggered by the
parent stream. When this callback is fired, the streamId associated with
the stream is no longer valid. In this commit, I start tracking the
registered streams and don't bother registering a subdirectory with the
watch service if one of its parents is already being monitored.