Reduce timeouts
These timeouts are all excessively long from what I can tell. This
change makes the whole test suite take ~10 seconds instead of ~45
seconds (the *WatchServiceSpec decline from ~45 seconds to ~2.5 seconds
on my machine). I found, however, that on travis, longer timeouts are
needed because the jvm needs to warm up and the timeouts are not short
enough in that scenario.