Set new file modified times in test
I was having a lot of trouble getting the PollingWatchServiceSpec to
pass on travis with osx and it turns out this is because the osx agents
are using HFS+ not APFS. The HFS+ file system doesn't support
millisecond precision on modified times. To fix this, set the file
modified time to be in the past when writing a new file.