Don't directly expose keys
The keysWithEvents was exposing unsynchronized mutable data. Even though
it was only accessed from a different thread inside of synchronized
blocks, it seems better to guarantee this behavior by adding a helper
accessor that ensures synchronization.

For binary compatibility had to provide a getter (that I deprecated) for
_keysWithEvents. I suspect that no one is using these apis outside of
the PollingWatchService file since PollingThread is private which
suggests that it might make sense to just add a mima filter instead.