Refactor EventMonitor
In subsequent commits I'm going to add support for a new EventMonitor
implementation that uses a file system cache. I factored out the
anti-entropy logic from the event thread so that I could share it with
the file cache implementation.

Bonus: I fixed all of the intellij warnings so that the file has a green
checkbox at the top