Refactor EventMonitor to use a FileEventMonitor
In previous commits, I add the Observable trait along with the
FileEventMonitor trait. These provide the fundamental building blocks
that can be used to implement the simpler EventMonitor trait. To
increase the generality of the provided methods, I add the
WatchServiceBackedObservable class. This implements the Observable
interface mostly using the implementation provided in the
EventMonitorImpl class. Because the implementation logic is largely the
same and the EventMonitor interface is easily expressed using the
FileEventMonitor interface, I was able to replace the EventMonitor apply
method implementation.

As a stand-alone commit, there is not much point to this change. The
WatchServiceBackedObservable is, however, a more useful class than the
EventMonitorImpl and will find use in subsequent commits that replace
the EventMonitor entirely.

One important implementation change though is that there is no longer a
separate thread for polling System.in. The EventMonitor
terminationCondition is now applied by polling the FileEventMonitor for
small periods (i.e. 10.milliseconds) and checking System.in between
polls. This not only reduces the bookkeeping associated with managing
that thread, but it also fixes a bug where in sbt the '~console' command
does not work because the user input thread will detect any input during
the console session (which is sure to include an <enter>). There is
still a background thread for polling the underlying WatchService for
events and invoking the observer callbacks when events are detected.