Add FileRepository to io
Throughout the sbt code base, it is necessary to list the contents of a
directory. For large projects, this can become slow because it requires
a lot of file I/O and also may perform computationally expensive work
like computing the sha1 sum of the file contents. It can then be a big
performance win to introduce an in-memory cache of the file system that
is updated as the underlying file contents change.

In this commit, I add three traits that will be used to optimize file
system access:

1) TypedPath -- represents a path on the file system, but may cache some
   of the attributes of the file, e.g. whether or not the file is a
   directory, so that filters may be applied without necessarily having
   to query the file system. Note that the equivalent apis in
   java.io.File or java.nio.file.Files must stat the file which is an
   expensive io operation.

2) FileTreeView -- this provides a simple api for listing the contents
   of a directory. Unlike, say, java.nio.file.Files.walk, the
   FileTreeView.list method returns TypedPath instances, allowing for
   fast filtering.

3) FileTreeDataView -- augments FileTreeView with a listEntries method
   that returns instances of Entry[T], which augment TypedPath with a
   data value. For example, the Entry.value may be the sha-1 sum of a
   source file or it could be the last modified time. Regardless, the
   idea is that it allows the implementation to optionally cache this
   value so that it does not need to do any file system operations to
   return the value to the user. I added this specifically so that I
   could add a zinc external hook in sbt that allows zinc to check which
   source files have changed without having to re-compute these values,
   which is generally bound by disk io, not cpu.

4) FileRepository -- augments FileTreeDataView with the option of
   registering paths for monitoring and adding observers that are
   notified whenever the monitored paths are updated. The repository
   specification does not require caching of the Entry values, though it
   is generally implied.

These traits mirror the
com.swoval.files.{ FileTreeView, FileTreeDataView, FileTreeRepository, TypedPath}
interfaces defined in the swoval file-tree-views library. I do not want
to directly expose the swoval apis to sbt users hence I define the
shadow api. Subsequent commits will provide implementations of these
interfaces.