Switch to Files.walkFileTree in PathFinder
The handleFileDescendant method was recursively constructing a list of
results by calling File.list and recursively calling back into itself
for each directory that it found in the listed files. There were two
performance bugs in this approach. One was that handleFileDescendant
actually called File.list twice for each directory. The other was that
after it got the second list of files, it had to filter the results
based on whether or not it was a directory. When I called jstack on my
sbt process during a slow test run, invariably one of the threads was
bloced on File.isDirectory.

The fix is simple, if verbose. Use Files.walkFileTree instead. It is
verbose, but it recursively lists all of the files in the tree specified
by the path. An alternative would be to use Files.walk. I didn't go this
route because Files.walk is somewhat hard to reason about because the
iterator can throw an exception at any time. This is a problem if the
directory is being traversed and modified at the same time. With
Files.walkFileTree, I know that if, for example, a directory disappears
from out from underneath us, then we'll just move on without throwing an
exception.