Treat ERROR_ACCESS_DENIED as file not found
After adding the .appveyor.yml, I found that my builds would typically
fail and usually one of the polling watch service threads died with an
IOException with message "CreateFile() failed with error 5." I tracked
this down to ERROR_ACCESS_DENIED, which seemed weird since any of the
files that were being polled were generated by the test. It turns out
that windows will return ERROR_ACCESS_DENIED when the file was in a
pending delete state.
https://stackoverflow.com/questions/6680491/why-does-windows-return-error-access-denied-when-i-try-to-open-a-delete-pended-f
While this feels slightly wrong, it doesn't seem completely unreasonable
to consider ERROR_ACCESS_DENIED to be like a FileNotFoundException. In
either the deleted or privileged file cases, the file is effectively not
findable.

An alternative would be to throw a different exception and catch that in
getModifiedTimeOrZero.