Add swoval implementation for (DescendendOrSelf|Child)PathFinder
The built in java apis that allow recursive file tree traversal suffer
from the need to stat every file to determine whether or not it is a
directory or a file. This is very expensive because stat-ing all of the
files can be much more expensive than retrieving them from the OS.
Fortunately, the widely use operating systems (OSX, Windows, Linux) all
provide apis for listing a path that provide limited information about
the returned files. Critically, they will indicate whether the file is a
directory or regular file (and on Linux and OSX, also whether or not the
file is a symbolic link). The built in java apis discard this
information and must make a stat (or windows equivalent) system call on
each returned path, which is very slow. Since there are many situations
where we do not need the other attributes returned by stat, this is a
very fruitful optimization.  Depending on the directory structure, I
have seen the swoval implementation take anywhere between 20% and 50% of
the time to run that the built in java implementations take to run. For
large projects this can be a difference between compilation taking 1
second to start, vs. 250 milliseconds, which is quite noticeable.

Because I have only tested on Windows (32/64 bit), OSX (64 bit) and
Linux (32/64 bit), I still provide an alternative nio based
implementation that can be toggled with the
"sbt.pathfinder.implementation" system property. This also guards
against the possibility that the swoval implementation does not work
correctly for some esoteric directory layouts. Fundamentally both
implementations use the same native os calls (as described above), so
the risk seems small but worth guarding against. The swoval project
itself also has a nio implementation of com.swoval.files.FileTreeViews
and I run tests on all platforms that the native and nio based
implementations return the same results.