Add swoval implementations of traits
Now that we have the trait definitions in FileRepository.scala, we need
implementations. I add two basic implementations for FileTreeView and
FileTreeRepository. A FileTreeDataView can be created using an extension
method that converts a FileTreeView into a FileTreeDataView using a
provided converter function.

The new classes DefaultFileTreeView and FileRepositoryImpl effectively
just delegate to an underlying swoval class and perform necessary
conversions between sbt.io and com.swoval.files data types. I also add
the FileRepositorySpec which tests some of the basic semantics of the
FileRepository. I have a significantly more comprehensive test suite of
the swoval implementation of its FileRepository interface that could
eventually be fully ported over to sbt if desired.