Bump swoval version
I need to upgrade the swoval library to provide a base implementation
for FileRepository. In the updated swoval version, I renamed the package
that contains the apple file events related code. More importantly, I
also added a java implementation of MacOSXWatchService that is
significantly more robustly tested than the scala implementation that
presently exists in io. To keep things working, I updated the io
implementation to just delegate to the swoval version.

The swoval file-tree-views package is implemented in pure java, which
is nice for io, which must build against a number of scala packages,
including 2.13.X -- a moving target.