Add HybridPollingFileRepository
There are some cases where it is not possible to monitor a directory
using the os file system notification hooks. For example, if a directory
is actually an nfs mount, this won't work on Linux (I'm not completely
sure under which conditions file system notifications don't work on the
various operating systems). To work around this, I add the
HybridPollingFileRepository which uses a FileRepositoryImpl to cache and
monitor some paths, but doesn't cache or monitor the rest. When listing
paths, if the path to be listed isn't required to be polled, we still
need to partition the results because it's possible that there are
subdirectories that require polling*.

We can make an event monitor for continuous build by combining the
repository event monitor with a polling watch service for the paths that
are not monitored.

* This overlapping use case is discouraged. If it is necessary to mix,
say, nfs mounted directories with local directories in the repository,
then the nfs mounted directories would ideally be in a separate
directory tree from local directories. For example, consider the
following direectory layout:

root (local):
  a (nfs)
  b (local)

Assuming that shouldPoll returns true for directory (a), it would be
fine to register root non-recursively and then to separately register a
and b. I am discouraging registering root recursively, although I make a
best effort at supporting that use case.

I have a hunch that #4179 can be resolved using the
HybridPollingFileRepository to monitor the project that is outside of
the home directory.