Join PollingThread on close
I noticed that the polling thread was only implicitly halted by setting
the close variable to false. While this did seem to work ok, I generally
think it's good practice to ensure that we join threads in close.
Otherwise, it's possible to have an unnoticed thread leak. With a 5
second join timeout, it will be noticeable if the thread is not quickly
joinable, but it won't halt the application.

Also, we only want to interrupt/join once, so I switched to an
AtomicBoolean for the closed state to use compareAndSet.