Add FileEventMonitor
In the initial implementation of EventMonitor, I was working with a
largely asynchronous mental model of the system. The EventMonitor was
effectively a black box synchronization point that allowed the user to
wait for anything to happen (but that gave no indication at all about
what happened). Moreover, I conflated two types of events in the
EventMonitor: user input and file system notifications. This conflation
made it somewhat unwieldy and difficult to work with. In this commit, I
add the FileEventMonitor trait which is solely responsible for managing
file system notifications.

While the api is fairly similar to EventMonitor, it has two key
differences:
1) It may be polled for a finite duration. This allows us to interleave
   other work in between polls. More specifically, in the sbt use case,
   we can remove the user input thread that exists in EventMonitor and
   check System.in between polling periods

2) When polled, it returns a sequence of events for every file that was
   updated since the last time it was polled or it blocks until new
   events are received (which are immediately returned) or it times out
   (returning an empty sequence). This allows the user more control over
   what exactly triggers a build. For example, if a file's last modified
   time is changed, but not the contents, the FileEventMonitor may
   return an Update event. The user can choose to ignore this event if
   the contents have not changed. This was not possible with the
   EventMonitor since it discarded all of the information about what
   event actually occurred.

The implementation uses an Observable to watch for events. This is
particularly convenient for tests because we can test the semantics of
the api without actually needing to modify the file system. As a general
rule, tests that actually modify the file system are harder to reason
about because it can be difficult to synchronize the state of the file
system with the state of the program.

A nice property of the new interface is that it is straightforward to
implement anti-entropy quarantining as a decorator on top of a more
simple FileEventMonitor. I uncovered a few tricky edge cases while
working on this, however. For one, it is very common for programs to
implement 'save' with a move of an internal file onto the target
location. This can manifest as a watch process seeing a delete
immediately followed by a creation. When this happens, depending on the
speed of updates, it is actually possible for the program to begin
handling the update before the new file exists. This can be particular
pernicious when paired with anti-entropy because the creation event may
be dropped leaving the program in an inconsistent state. To fix this, I
quarantine ALL deletions for a small period of time. They are removed
from quarantine either when a creation is detected or the quarantine
period expires. That is the logic behind the quarantinedEvents map in
AntiEntropyFileEventMonitor.