Add Observers
This class will allow us to selectively add and remove Observers from a
given Observable. The motivation is that a user of a FileRepository may
only want to receive file notification events when their program is in a
certain state. This class also can function as a bus between an Observer
and an Observable allowing for dynamically updating Observers even for
Observables that do not directly support adding and removing Observables
(but perhaps take an Observable as a constructor argument).