Remove references to EventMonitor
Now that I've replaced the functionality in EventMonitor with
FileEventMonitor (which has implementations using FileRepository and
WatchServiceBackedObservable), I am replacing the tests of the
EventMonitor interface with tests of the FileEventMonitor interface. The
capabilities that were previously tested all still exist, so no test
cases are created or removed in this commit.