Deprecate EventMonitor
I'd like to remove EventMonitor entirely, now that I've replaced its
functionality with FileEventMonitor. I'm actually not clear on whether
or not it is kosher to remove this trait right now. It has always only
had private[sbt] visibility. If, however, an old version of sbt > 1.1.4
is run with a new version of io, the program would crash. I'm not sure
if that's possible though because it seems that sbt requires that it is
run against the version of io that was used to compile it. If that is
true, then this commit could just delete EventMonitor.scala, because it
is only directly referenced in this file.