Rewrite PollingWatchService
The PollingWatchServiceSpec was frequently failing both in CI and on my
mac. I found it fairly difficult to reason about because of the mixture
of mutable and immutable data structures so I rewrote the entire thing
from scratch. I made a few significant design decisions:

1) I consolidated the keys and watched maps into a single
   ConcurrentHashMap, registered. To do this, I embed the accepted event
   types in the PollingWatchKey.

2) I rework things to hold the lock for as little as possible. One of
   the reasons for test failures is that both register and the polling
   thread would need the lock and would block each other. Now, we only
   hold the lock when we update the contents of the lastTimestamps map.

3) The PollingThread has no direct state or methods. Since it's an inner
   class of the PollingWatchService, it needs neither.

4) I batch offer the events. One of the issues that I was running into
   was that sometimes a single event would get dropped. This seemed to
   be due to a race condition where the polling thread would be offering
   multiple events for a key while the main thread would be polling the
   service. In some rare events, it was possible for an event to be
   dropped. Now, I pre-compute all of the events, batch add them to the
   key and then batch add all of the keys.

After making these changes, my spurious failures seem to have
disappeared. I also went back and re-worked my new version so that the
layout was similar to the old one. I removed a number of methods that
were only in private classes which required updating the mima filters
since mima does not seem to work how I'd expect with private classes.