Fix flaky test
This test occasionally fails on travis because the expected last
modified times don't match. This is likely because sometimes the
callback will fire before the file's last modified time has been
updated. To fix this, ensure that we only trigger the latch when the
file's last modified time has been updated to the expected value.