Make IO.classfileLocation consistent
I have been having trouble getting the sbt tests to pass (both scripted
and runProj/test) since #185. The big problem that I've been running
into is that for some classes IO.classfileLocation returns a full
absolute path to the class file* while for others it returns just the
base directory of the class file. In this commit, I rework
classfileLocation to just turn the base directory (or jar file) for the
class. It is then up to the user to append the class name if necessary.
A separate PR in zinc is necessary to adapt this logic.

* There are effectively two cases: class files in a directory and class
files in a jar. The former should be url with the format:
file:$ABSOLUTE_PATH_TO_CLASS_FILE
while the latter shoud have the format:
jar:file:$ABSOLUTE_PATH_TO_JAR!/${RELATIVE_CLASS_FILE_NAME}