Add base directory to swoval pathfinder implementation
Somehow much of sbt seemed to work on mac and linux despite my
inadvertently changing the semantics of io '**' to not include the
directory that is recursively listed. It did, however, cause failures on
windows. At any rate, I fixed the issue by adding the base directory to
the result set in DescendantOrSelfPathFinder.default. To ensure that the
semantics of the two PathFinder implementations remain the same, I
rework PathFinderSpec so that it tests both implementations.

Bonus: add some params to scaladocs