Use nio by default on windows
Although the swoval implementation works fine on my vm, I still have
not been able to get the appveyor builds to pass when it is used. We'll
just use the previous implementation until I can get appveyor working
with it.

Note that I did get a personal appveyor build of sbt to work when run
against this io commit.