Fix anti-entropy bug
I noticed that sometimes my builds against the latest sbt nightly would
not re-trigger if my source file was touched while the test was running.
It turned out that the reason was that the condition that checked if the
file event occurred within the anti-entropy period was incorrect. I'm
not really sure why I wrote the initial condition this way. I also renamed
recentEvents to antiEntropyDeadlines to make the code easier to read.

The new test fails without the change to FileEventMonitor.scala.