Add extension filter
It is very common in sbt to want to filter files with a particular
extension, or files whose extension is in some set of extensions. I
introduce the ExtensionFilter to special case this. I don't really
expect users to directly make an ExtensionFilter, but now when a user
creates a filter like:
watchSources in Compile += Source(sourceDirectory.value, "*.scala" | "*.java", NothingFilter)
the Source will actually be create with an ExtensionFilter. The
advantage of this is that regex matching is slow, whereas finding the
extension of a file is very fast. Once we have the extension, we can
just check if it is in the set of accepted extensions.