Add Glob trait to io
There are a number of places within sbt where we want to refer to a
file tree pattern. This is useful when we want to watch or cache a
portion of the file system tree. We already have a similar abstraction,
sbt.internal.io.Source, but there are a few issues with this:

1) the name is confusing. It represents a collection of files but the
   name makes it sound like it corresponds to a single source file.
2) it contains two filters, but it really only needs one for the
   common use cases.
3) there is no support ecosystem around Source to make it easy to use
   them to, say, actually retrieve the list of files described by the
   pattern. Nor is there a dsl for specifying them.

We could solve #3 by just building out that ecosystem, but I couldn't
think of a good reason to further extend an internal class that already
has some semantic issues.

My new approach is to introduce the new trait Glob. It is a sealed trait
with a minimal api. While I could have made it a final class, using a
sealed trait with no default implementations felt slightly more flexibly
for future api changes.

To facilitate using Glob with the PathFinder dsl, I introduce an
extension class for File called GlobBuilder whose api mirrors a subset
of the PathFinder api except that the methods return Glob instances
rather than PathFinder instances. I further provide an implicit
conversion from Glob to PathFinder. Combining these two, we should be
able to preserve source compatibility with the existing PathFinder dsl
while slightly changing the semantic meaning.