Add FileFilters for Glob and Traversable[Glob]
One of the key use cases for Glob is to be able to filter files on the
basis of whether or not the Glob covers the file path. For a single
glob, I add an extension method, toFilter, to GlobOps. This creates a
FileFilter that validates whether or not the file is described by the
glob pattern. I also add an extension class for Traversable[Glob] that
provides a toFilter method that converts all of the Globs in the input
collection to a filter (using the GlobOps.toFilter method) and accepts
the file if any of the globs accept it.

In both case, I wanted the conversion to FileFilter to be explicit
because implicit conversions are generally more confusing than extension
methods.