Add alternative antiEntropy method
In overhauling the watch implementation in sbt, I realized that it's
convenient to be able to wrap a FileEventMonitor in an
AntiEntropyFileEventMonitor. This makes it easier to aggregate a
collection of FileEventMonitor into a single FileEventMonitor that can
reject redundant events that come from multiple FileEventMonitor
instances if more than one monitor is observing the same portion of the
file system tree.

I also removed the default value from the existing antiEntropy method
because methods with default values cannot be evolved in a binary
compatible way. If there really is a need for the default value, then we
can add telescoping methods.

Finally, I added the retentionPeriod parameter to
FileEventMonitor.antiEntropy. The motivation was that I would sometimes
see spurious triggers in sbt continuous builds if the build itself took
a while. This was because there was a long period between calls to
FileEventMonitor.poll. During that time, we accumulated a number of
events that may have occurred a while ago in wall time.