Replace Source based apis with Glob based apis
This commit deprecates Source and replaces all of it's uses with apis
that use Globs. No old behavior should change as a result of this
refactoring.

The list/listEntries/register methods of FileTreeRepository were a bit
unwieldy since they required 2 or 3 parameters. I simplified this by
adding extension methods to the appropriate traits that add variants of
these methods that take a Glob as input.