Add conversion function from TypedPath to File
I have inlined this logic throughout io and sbt so it makes sense to
just make it available as an extension method on TypedPath. I add it as
an extension method to give us the flexibility to remove* this or update
semantics without breaking binary compatibility.

I mean remove in the sense of having it automatically available without
an import.