Add unsealed IOSyntax trait
There are a number of redundant implicit conversions defined in the
AllSyntax.scala file in sbt with a TODO suggesting that the defs be
shared with IO. I had to inherit IOSyntax1 rather than IOSyntax0 because
sbt defined an alternative "Alternative" class which led to some binary
compatibility issues. I don't think that Alternative should make it to
sbt 2 since it's barely used and having two versions of the same class
in the sbt project is problematic.