Bump swoval
I found an off by one error that would erroneously evict a subdirectory
from monitoring if both the subdirectory and its parent were registered
with depth=0.  This is fairly rare in normal sbt builds since usually
only the base directory is registered with depth=0, but I discovered it
in a scripted test and it surely would have caused an issue for some
users.