Add Registerable
In sbt I want to make it possible to modify the list of registered paths
with a WatchServiceBackedObservable to provide an alternative file
monitoring implementation if for any reason users are having issues with
the FileTreeRepository implementation. To unify the api, I add the
Registerable trait which provides the register/unregister methods that
were previously defined in FileTreeRepository.

After defining the Registerable trait, I implemented it