Add LegacyFileTreeRepository
This adds an alternative FileTreeRepository implementation that uses a
WatchService for file event monitoring but polls the file system
whenever the user wants to list the contents of a directory. The
motivation is that in sbt I want to be able to use this as an
alternative FileTreeRepository for users who have issues with the
caching file tree repository.

I added the LegacyFileTreeRepositoryEventMonitorSpec which ensures that
this new repository satisfies the same spec as the other ones.