Change Glob representation to path
I think that moving forward, it makes sense for the internal data
structures of sbt to use Path rather than File for the representation.
To smooth this transition, I rework Glob to be represented by a Path
rather than a File and a filter that is just a function from
TypedPath => Boolean. I do, however, leave the File based api intact so
that Globs can still be used seamlessly in existing sbt builds, which
predominantly use File instances (e.g. Keys.baseDirectory).